Consolidate logaddexp definitions.
